wip organize tests to eash pass's Test.hs.
