+++ plutus-core/plutus-ir/test/PlutusIR/Transform/Beta/Tests.hs
@@ -6,8 +6,11 @@ import Test.Tasty.Extras
 import PlutusCore qualified as PLC
 import PlutusCore.Quote
 import PlutusIR.Parser
+import PlutusIR.Properties.Typecheck
 import PlutusIR.Test
 import PlutusIR.Transform.Beta qualified as Beta
+import Test.QuickCheck.Property (withMaxSuccess)
+import Test.Tasty.QuickCheck (testProperty)
 
 test_beta :: TestTree
 test_beta = runTestNestedIn ["plutus-ir/test/PlutusIR/Transform"] $
@@ -20,3 +23,7 @@ test_beta = runTestNestedIn ["plutus-ir/test/PlutusIR/Transform"] $
             , "multiapp"
             , "multilet"
             ]
+
+test_typecheck :: TestTree
+test_typecheck = testProperty "typechecking" $
+    withMaxSuccess 3000 (pure_typecheck_prop Beta.beta)
