+++ plutus-core/plutus-ir/test/PlutusIR/Transform/CommuteFnWithConst/Tests.hs
@@ -7,6 +7,10 @@ import PlutusIR.Parser
 import PlutusIR.Test
 import PlutusIR.Transform.CommuteFnWithConst qualified as CommuteFnWithConst
 
+import PlutusIR.Properties.Typecheck
+import Test.QuickCheck.Property (withMaxSuccess)
+import Test.Tasty.QuickCheck (testProperty)
+
 test_commuteDefaultFun :: TestTree
 test_commuteDefaultFun = runTestNestedIn ["plutus-ir/test/PlutusIR/Transform"] $
     testNested "CommuteFnWithConst" $
@@ -17,3 +21,7 @@ test_commuteDefaultFun = runTestNestedIn ["plutus-ir/test/PlutusIR/Transform"] $
         , "multiplyInt" -- this tests that the function works on multiplyInteger
         , "let" -- this tests that it works in the subterms
         ]
+
+test_typecheck :: TestTree
+test_typecheck = testProperty "typechecking" $
+      withMaxSuccess 3000 (pure_typecheck_prop CommuteFnWithConst.commuteFnWithConst)
